requirements.txt
Flask==2.0.1
Werkzeug==2.0.3          # Flask 2.0.1 ile daha uyumlu
Jinja2==3.0.3            # Flask 2.0.1 ile daha uyumlu
MarkupSafe==2.0.1        # Jinja2 ile uyumlu
itsdangerous==2.0.1      # Flask 2.0.1 ile uyumlu
click==8.0.4             # Flask 2.0.1 ile uyumlu

tensorflow==2.14.0
keras==2.14.0            # TensorFlow ile uyumlu Keras
numpy==1.24.3            # TF 2.14 ile uyumlu NumPy

Pillow==9.0.0            # Pillow 10 değil, 9 kullan (TF image ile daha uyumlu)
scikit-learn==1.0.2
opencv-python==4.5.3.56
